Refactor: inline components related to AccountSelection
