+++ src/hooks/useBrowserIcon.js
@@ -1,20 +1,18 @@
 import React, { useMemo } from 'react';
 import styled from 'styled-components';
 
-import useLanguage from 'hooks/useLanguage';
 import { ReactComponent as MetaMaskLogo } from 'images/metamask.svg';
 import { ReactComponent as TrustLogo } from 'images/trust-logo.svg';
 import coinbaseWalletLogo from 'images/coinbase-wallet.png';
 import alphaWalletLogo from 'images/alpha-wallet-logo.png';
 import { wallets } from 'utils/web3';
 import ImTokenLogo from 'components/ImTokenLogo';
-import IconButton from 'components/IconButton';
 
 const MMLogo = styled(MetaMaskLogo)`
   margin: -5px 0;
 `;
 
-export const useBrowserIcon = provider =>
+const useBrowserIcon = provider =>
   useMemo(() => {
     if (provider === wallets.METAMASK) {
       return <MMLogo />;
@@ -37,12 +35,4 @@ export const useBrowserIcon = provider =>
     }
   }, [provider]);
 
-export default function BrowserProviderButton({ provider, ...props }) {
-  const { lang } = useLanguage();
-  const icon = useBrowserIcon(provider);
-  return (
-    <IconButton icon={icon} {...props}>
-      {lang.providers[provider] || 'Active Wallet'}
-    </IconButton>
-  );
-}
+export default useBrowserIcon;
